requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask==2.3.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
pydantic==1.10.12
PyJWT==2.8.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.41.1
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
typing_extensions==4.7.1
Werkzeug==2.3.7
